CWE-172179: In CommonConstants added new Constant CERT_PARM = "-certificate"
In this jira created 4 common routines:

1) CpwrGlobalConfiguration.encodeX509Certificate(X509Certificate
x509Certificate) that return a Base64 encoded representation of the
X509Certificate or null if no X509Certificate provided.

2) CpwrGlobalConfiguration.encodeX509Certificate(byte[]
byteX509Certificate) that encode the specified X509Certificate specified
as an
array of bytes as a Base64 string.

3)  CpwrGlobalConfiguration.getCertificate(StandardCertificateCredentials
credentials) this method returns a string of Base64 encoded
representation of the X509Certificate if we found this certificate in
the certificateChain using the keystore or it will return null  if no
certificate  founded in the certificateChain.

4) CpwrGlobalConfiguration.ArgumentListBuilder() this method builds the
list of arguments to pass to the CLI. We can use this method to complete
the work in jira CWE-172871.

In HostConnectionTest created the getCertificateStr() to test
globalConfig.getCertificate() using certificate of type (.pfx ).
We need to use a certificate of type (.pfx) because we use a default
keystore of type "PKCS12".

The problem is we have to export the private Key in the certificate to
make it of type ".pfx" and we can't add this certificate in our plugin
for security since we publish this code in GitHub so I commented out
getCertificateStr().

We can't run getCertificateStr() as part of HostConnectionTest since I
didn't add the certificate that we need to use in the test but we can
run this test manually after creating a certificate of type (.pfx ) this
link shows how to do this
(https://mindfulsoftware.com.au/blog/1409024/Converting-a-cer-file-to-pfx-using-the-Windows-MMC-snapin)
then add this certificate to compuware-common-configuration-plugin ->
src/main/resources.